requirements.txt
fastapi
uvicorn
pydantic
python-jose[cryptography]
passlib[bcrypt]==1.7.4
bcrypt==3.2.0
pyjwt
httpx
PyYAML
pytest
pytest-asyncio
requests
aiosqlite
google-generativeai
python-dotenv
langchain-core
langchain-groq

langchain
langchain-community
langchain-huggingface
faiss-cpu
sentence-transformers
pypdf
unstructured
reportlab
requests
python-multipart